# mobile/openapi/identity-service.yaml
...
openapi: 3.1.0
info:
  title: TD IT Solutions — Identity Service API
  version: "1.0.0"
  description: >
    Feature 001 (Customer Account Creation & Authentication). Mediates all
    client access to Supabase Auth per ADR-0002's mediation principle —
    no client ever holds a Supabase-honoured credential (see api-design.md §1).
servers:
  - url: /v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Backend-minted access token (api-design.md §1). Never a Supabase-issued JWT.
    internalServiceAuth:
      type: apiKey
      in: header
      name: X-Internal-Service-Key
      description: Service-to-service auth for the internal-only account-status surface (§2.3 row 4). Not reachable by any client.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Client-generated UUID v4. See api-design.md §4.
...
  headers:
...
    RetryAfter:
      schema: { type: integer, description: Seconds. }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            requestId: { type: string, format: uuid }
...
    UserType:
      type: string
      enum: [customer, admin, security_company_operator, support_agent]
    AccountState:
      type: string
      enum: [pending_verification, active, suspended, deactivated]
    Account:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        userType: { $ref: '#/components/schemas/UserType' }
        accountState: { $ref: '#/components/schemas/AccountState' }
        mfaRequired: { type: boolean }
        mfaEnrolled: { type: boolean }
        partnerOrganizationId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
    SessionTokens:
      type: object
      required: [accessToken, refreshToken, expiresIn, sessionId]
      properties:
        accessToken: { type: string, description: Backend-minted JWT, api-design.md §1. }
        refreshToken: { type: string }
        expiresIn: { type: integer, example: 600, description: Seconds (10 min ceiling, api-design.md §2). }
        sessionId: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Missing/invalid/expired/revoked bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    UpstreamUnavailable:
      description: Supabase Admin API unreachable/timed out (api-design.md §6).
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /mfa/enroll:
    post:
      operationId: mfaEnroll
      summary: Starts TOTP enrollment. Callable by an unauthenticated-but-invitation-verified privileged user (first login) or an authenticated customer opting in.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                accountId:
                  type: string
                  format: uuid
                  description: Required only for the forced-first-login (post-accept, pre-session) path; omitted when called with a bearer token.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  qrCodeImage: { type: string, description: Base64-encoded PNG. }
                  manualEntryKey: { type: string }
                  enrollmentId: { type: string, format: uuid }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }
...
  /session/logout:
    post:
      operationId: logout
      responses:
        '204':
          description: Current session revoked immediately (api-design.md §2.1, Mechanism 1).
        '401': { $ref: '#/components/responses/Unauthorized' }
  /session/logout-all:
    post:
      operationId: logoutAll
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: All sessions for this account revoked immediately.
        '401': { $ref: '#/components/responses/Unauthorized' }
  /session/refresh:
    post:
      operationId: refreshSession
      summary: >
        Exchanges a refresh token for a new access token. This is the
        Mechanism-2 chokepoint (api-design.md §2.3) — always performs a live
        account_status_cache read before minting a new access token. Also
        (v1.1.0, §11 Amendment C, ratifying mobile-app-foundation/
        architecture.md §3.2's M-01) the device-consistency chokepoint for
        FR-20 device-binding: if `deviceId` is presented and does not match
        the session's stored device, the refresh is rejected and the entire
        rotation family plus every other active session for the account is
        revoked (`device_mismatch` — treated at least as seriously as
        refresh-token-rotation reuse, same blast-radius response as §3.1's
        reuse detection).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
                deviceId:
                  type: string
                  nullable: true
                  description: >
                    Added v1.1.0. Optional, additive — omitting it (or
                    sending null) does not itself trigger a mismatch; it
                    simply means this refresh isn't checked against the
                    session's stored device. Mobile clients SHOULD send the
                    same `deviceId` presented at login for the device-
                    mismatch protection to actually engage.
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTokens' }
        '401':
          description: >
            Refresh token invalid, expired, revoked, reuse-detected
            (rotation chain break — api-design.md §3.1), or (v1.1.0)
            device-mismatch detected — identical status code and error
            envelope for all four; the client cannot and must not
            distinguish them beyond forcing a full local logout in every
            case (mobile-app-foundation/architecture.md §2.4).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '423':
          description: Account is suspended/deactivated — refusing to mint a new access token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503': { $ref: '#/components/responses/UpstreamUnavailable' }

  # ---------------------------------------------------------------
  # Password reset
  # ---------------------------------------------------------------
...
  /account/me:
    get:
      operationId: getCurrentAccount
      summary: >
        Always a live account_status_cache read (api-design.md §2.3) — never
        served from the JWT claim alone. Drives client-side BR-2 gating UI.
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ---------------------------------------------------------------
  # Internal-only surface (server-to-server, not client-facing)
  # ---------------------------------------------------------------
...